Moving sample() from QuestionDefinitionBuilder to ApplicantQuestionRendererFactory
