+++ universal-application-tool-0.0.1/app/views/questiontypes/ApplicantQuestionRendererFactory.java
@@ -1,17 +1,24 @@
 package views.questiontypes;
 
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
+import models.LifecycleStage;
+import services.Path;
 import services.applicant.ApplicantData;
 import services.applicant.question.ApplicantQuestion;
+import services.question.QuestionOption;
 import services.question.exceptions.UnsupportedQuestionTypeException;
 import services.question.types.QuestionDefinition;
 import services.question.types.QuestionDefinitionBuilder;
 import services.question.types.QuestionType;
 
+import java.util.Locale;
+
 public class ApplicantQuestionRendererFactory {
 
   public ApplicantQuestionRenderer getSampleRenderer(QuestionType questionType)
       throws UnsupportedQuestionTypeException {
-    QuestionDefinition questionDefinition = QuestionDefinitionBuilder.sample(questionType).build();
+    QuestionDefinition questionDefinition = sample(questionType).build();
     ApplicantQuestion applicantQuestion =
         new ApplicantQuestion(questionDefinition, new ApplicantData());
     return getRenderer(applicantQuestion);
@@ -41,5 +48,29 @@ public class ApplicantQuestionRendererFactory {
         throw new UnsupportedOperationException(
             "Unrecognized question type: " + question.getType());
     }
+  }
+
+  public static QuestionDefinitionBuilder sample() {
+    return sample(QuestionType.TEXT);
+  }
+
+  public static QuestionDefinitionBuilder sample(QuestionType questionType) {
+    QuestionDefinitionBuilder builder =
+        new QuestionDefinitionBuilder()
+            .setName("")
+            .setDescription("")
+            .setPath(Path.create("sample.question.path"))
+            .setQuestionText(ImmutableMap.of(Locale.US, "Sample question text"))
+            .setQuestionHelpText(ImmutableMap.of(Locale.US, "Sample question help text"))
+            .setLifecycleStage(LifecycleStage.ACTIVE)
+            .setQuestionType(questionType);
+
+    if (questionType.isMultiOptionType()) {
+      builder.setQuestionOptions(
+          ImmutableList.of(
+              QuestionOption.create(1L, ImmutableMap.of(Locale.US, "Sample question option"))));
+    }
+
+    return builder;
   }
 }
